+++ backend/routes/ebay/ebay.js
@@ -1,7 +1,7 @@
 import express from "express";
 import axios from "axios";
-import { createSearchQuery } from "../../helpers/getMangaHelpers.js";
-import { getEbayToken } from "../../helpers/eBayTokenHelpers.js";
+import { createSearchQuery } from "../../helpers/ebay_helpers/ebay_get_data_helpers.js";
+import { getEbayToken } from "./ebay_token.js";
 
 const router = express.Router();
