+++ frontend/src/components/result_item/isbndb_item.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { useTheme } from '@mui/material/styles';
 import Box from '@mui/material/Box';
 import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
@@ -8,7 +7,6 @@ import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
 
 const IsbnDbItem = (props) => {
-    // const theme = useTheme();
 
     const onClickAdd = () => {
         props.onAdd(props.item);
