+++ frontend/src/helpers/db_helpers/authorHelpers.js
@@ -1,8 +1,9 @@
-import { getData, postData } from "./apiHelpers";
+import { getData, postData } from "../apiHelpers";
 
 // Change Later: Multiple names or people with more than just their firstname/surname are given to splitName
 
 export function splitName (name) {
+    
     let splitter = name.includes(', ') ? ', ' : name.includes('. ') ? '. ' : name.includes(' ') ? ' ' : undefined;
 
     // Get authors first and last name in respective variables
@@ -32,6 +33,7 @@ export async function addAuthor (name) {
 }
 
 export function isAuthorEqual (givenName, itemName) {
+
     let splitGiven = splitName(givenName)
     let splitItem = splitName(itemName)
 
@@ -39,6 +41,7 @@ export function isAuthorEqual (givenName, itemName) {
 }
 
 export async function authorExists (name) {
+
     const authors = await getData("authors")
 
     const filteredAuthorData = authors.filter(item => isAuthorEqual(name, item.name))
@@ -46,7 +49,25 @@ export async function authorExists (name) {
     return filteredAuthorData[0] !== undefined ? filteredAuthorData[0]._id : false; 
 }
 
+export async function getOrAddAuthors (authors) {
+
+    const author_ids = []
+    
+    for (const author of authors) {
+
+        let author_id = await authorExists(author);
+
+        // if author is not in manga db
+        if (!author_id) author_id = await addAuthor(author)
+        
+        author_ids.push(author_id);
+    }
+
+    return author_ids;
+}
+
 function arraysEqual (a, b) {
+
     if (a === b) return true;
     if (a == null || b == null) return false;
     if (a.length !== b.length) return false;
@@ -54,5 +75,6 @@ function arraysEqual (a, b) {
     for (var i = 0; i < a.length; ++i) {
       if (a[i] !== b[i]) return false;
     }
+
     return true;
   }
