+++ frontend/src/helpers/isbndb_helpers/isbnDataHelpers.js
