+++ frontend/src/pages/AddMangaPage.js
@@ -1,18 +1,15 @@
 import React, { useState } from 'react';
-import { Button, FormControl, TextField, Grid, Box } from '@mui/material';
-import Paper from '@mui/material/Paper';
-import { styled } from '@mui/material/styles';
-import { baseUrl } from '../config';
-import IsbnDbItem from '../components/result_item/isbndb_item';
-import { authorExists, addAuthor } from '../helpers/authorHelpers';
-import { seriesExists, addSeries } from '../helpers/seriesHelpers';
-import { mangaExists, addManga } from '../helpers/mangaHelpers';
-import { filterISBNData } from '../helpers/filterISBNData';
 
-// const Item = styled(Paper)(({ theme }) => ({
-//     padding: theme.spacing(1),
-//     textAlign: 'center',
-// }));
+import { Button, FormControl, TextField, Grid, Box } from '@mui/material';
+
+import { baseUrl } from '../config';
+
+import IsbnDbItem from '../components/result_item/isbndb_item';
+
+import { filterISBNData } from '../helpers/isbndb_helpers/isbnDataHelpers';
+import { getOrAddAuthors } from '../helpers/db_helpers/authorHelpers';
+import { getOrAddSeries } from '../helpers/db_helpers/seriesHelpers';
+import { getOrAddManga } from '../helpers/db_helpers/mangaHelpers';
 
 const AddMultipleMangaPage = () => {
     const [query, setQuery] = useState({
@@ -21,6 +18,7 @@ const AddMultipleMangaPage = () => {
         publisher: "",
         volume: "1"
     })
+
     const [data, setData] = useState([]);
 
     const handleChange = (e) => {
@@ -28,6 +26,7 @@ const AddMultipleMangaPage = () => {
     };
 
     const getBooks = async () => {
+
         const series = query.series;
         const author = query.author;
         const publisher = query.publisher;
@@ -41,18 +40,18 @@ const AddMultipleMangaPage = () => {
             },
             body: JSON.stringify(body),
         })
+
         response = await response.json();
 
         console.log(response);
-        // console.log(volume)
 
         const filteredResponse = filterISBNData(response, series, volume)
-
-        // console.log(filteredResponse);                                
+                            
         setData(filteredResponse)
     }
 
     const onAdd = async (manga) => {
+
         const authors = manga.authors;
         const series = query.series;
         const volume = query.volume;
@@ -60,44 +59,9 @@ const AddMultipleMangaPage = () => {
 
         const author_ids = await getOrAddAuthors(authors);
         const series_id = await getOrAddSeries(series, author_ids);
-        const manga_id = await getOrAddManga(manga, series_id, author_ids, volume, price)
+        const manga_id = await getOrAddManga(manga, series_id, author_ids, volume, price);
 
-        // let new_manga = {...manga, series_id: series_id, author_ids: author_ids, volume: volume, volume_price: volume_price}
-
-        // console.log(author_ids);
-        // console.log(series_id);
-        // console.log(manga_id);
-        
-    }
-
-    const getOrAddManga = async (manga, series_id, author_ids, volume, price) => {
-        let updated_manga = {...manga, series_id: series_id, author_ids: author_ids, volume: volume, price: price}
-
-        let manga_id = await mangaExists(updated_manga)
-
-        return manga_id ? manga_id : await addManga(updated_manga);
-    }
-
-    const getOrAddSeries = async (series, author_ids) => {
-        let series_id = await seriesExists(series);
-        
-        return series_id ? series_id : await addSeries(series, author_ids)
-    }
-
-    const getOrAddAuthors = async (authors) => {
-        const author_ids = []
-        
-        for (const author of authors) {
-
-            let author_id = await authorExists(author);
-
-            // if author is not in manga db
-            if (!author_id) author_id = await addAuthor(author)
-            
-            author_ids.push(author_id);
-        }
-
-        return author_ids;
+        return manga_id;
     }
 
     return (
@@ -105,7 +69,6 @@ const AddMultipleMangaPage = () => {
             <Grid container spacing={2}>
                 <Grid xs={3} sx={{ m: 3, textAlign: 'center' }}>
                     <FormControl>
-                            {/* <ElementSelect DropDown={SeriesDropDown} AddForm={AddSeriesForm} addText={"Add Series"} /> */}
                         <TextField 
                             sx={{ m: 1, minWidth: 240 }} 
                             label="Series" 
@@ -136,9 +99,7 @@ const AddMultipleMangaPage = () => {
                             onChange={handleChange} 
                         />
                         <Button  sx={{ m: 1, minWidth: 60 }} variant='contained' onClick={getBooks}>Get Books</Button>
-                        {/* <Button  sx={{ m: 1, minWidth: 60 }} variant='contained' onClick={getGantz}>Get Gantz</Button> */}
                     </FormControl>
-                    
                 </Grid>
                 <Grid xs={7} sx={{ mt: 3 }} textAlign={ "left" }>
                         {data.map(item => (
