--- frontend/src/pages/SeriesPage.js
@@ -1,1 +1,0 @@
-import React, { useState, useEffect } from 'react';
